fix(quality): resolve sonar issues

+++ modules/tools/pathfinder-sim/src/main/java/fr/riege/ebsl/tools/pathfindersim/world/minecraft/MinecraftWorldScenarioFactory.java
@@ -276,9 +276,9 @@ public final class MinecraftWorldScenarioFactory {
     }
 
     private static double squaredDistance(int ax, int ay, int az, int bx, int by, int bz) {
-        double dx = ax - bx;
-        double dy = ay - by;
-        double dz = az - bz;
+        double dx = (double) ax - bx;
+        double dy = (double) ay - by;
+        double dz = (double) az - bz;
         return dx * dx + dy * dy + dz * dz;
     }
